refactor(networkpath): migrate npcollector component to v2
Move from flat npcollectorimpl layout to v2 structure:
- comp/networkpath/npcollector/def/component.go (interface)
- comp/networkpath/npcollector/fx/fx.go (fx module)
- comp/networkpath/npcollector/mock/mock.go (test mock)
- comp/networkpath/npcollector/impl/ (implementation, moved from npcollectorimpl/)
Update comp/networkpath/bundle.go to use npcollectorfx.Module().
