+++ comp/networkpath/bundle.go
@@ -9,7 +9,7 @@ package networkpath
 import (
 	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 
-	"github.com/DataDog/datadog-agent/comp/networkpath/npcollector/npcollectorimpl"
+	npcollectorfx "github.com/DataDog/datadog-agent/comp/networkpath/npcollector/fx"
 )
 
 // team: network-path
@@ -17,6 +17,6 @@ import (
 // Bundle defines the fx options for this bundle.
 func Bundle() fxutil.BundleOptions {
 	return fxutil.Bundle(
-		npcollectorimpl.Module(),
+		npcollectorfx.Module(),
 	)
 }
