+++ comp/networkpath/npcollector/def/component.go
