+++ comp/networkpath/npcollector/impl/connfilter/connfilter.go
