+++ comp/networkpath/npcollector/impl/connfilter/connfilter_test.go
