+++ comp/networkpath/npcollector/impl/connfilter/connfiltertype.go
