+++ comp/networkpath/npcollector/impl/connfilter/defaultconnfilters.go
