+++ comp/networkpath/npcollector/impl/connfilter/testutils_test.go
