+++ comp/networkpath/npcollector/impl/connfilter/utils.go
