+++ comp/networkpath/npcollector/impl/npcollector_mock.go
@@ -10,25 +10,15 @@ package npcollectorimpl
 import (
 	"iter"
 
-	"go.uber.org/fx"
-
 	npmodel "github.com/DataDog/datadog-agent/comp/networkpath/npcollector/model"
-	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 )
-
-// MockModule defines the fx options for the mock component.
-func MockModule() fxutil.Module {
-	return fxutil.Component(
-		fx.Provide(NewMock),
-	)
-}
 
 type npCollectorMock struct{}
 
 func (s *npCollectorMock) ScheduleNetworkPathTests(_conns iter.Seq[npmodel.NetworkPathConnection]) {}
 
+// NewMock creates a mock npcollector component.
 func NewMock() Provides {
-	// Mock initialization
 	return Provides{
 		Comp: &npCollectorMock{},
 	}
