+++ comp/networkpath/npcollector/impl/pathteststore/pathteststore.go
@@ -14,7 +14,7 @@ import (
 	"golang.org/x/time/rate"
 
 	log "github.com/DataDog/datadog-agent/comp/core/log/def"
-	"github.com/DataDog/datadog-agent/comp/networkpath/npcollector/npcollectorimpl/common"
+	"github.com/DataDog/datadog-agent/comp/networkpath/npcollector/impl/common"
 )
 
 const (
