+++ comp/networkpath/npcollector/impl/utils.go
